requirements.txt
SpeechRecognition>=3.10.0
pyttsx3>=2.90
PyAudio>=0.2.13
pyautogui>=0.9.54
psutil>=5.9.0
groq>=0.9.0
google-genai>=0.8.0
pyperclip>=1.8.0
opencv-python>=4.8.0
face-recognition>=1.3.0
numpy>=1.24.0

# ── NEW: Screen Vision ─────────────────────────────────────────────────
google-generativeai>=0.7.0   # Gemini Vision API (free tier)
Pillow>=10.0.0                # Screenshot processing

# ── NEW: Audio Capture ─────────────────────────────────────────────────
pyaudiowpatch>=0.2.12         # Windows WASAPI loopback (system audio)
openai-whisper>=20231117      # Local speech-to-text (free, no API key)
# NOTE: openai-whisper also requires ffmpeg installed on PATH
# Download: https://ffmpeg.org/download.html → add the bin/ folder to PATH